feat:Lecturers fix:Some naming issues fix:removed some files
