+++ src/App.js
@@ -1,29 +1,29 @@
 import "bootstrap/dist/css/bootstrap.min.css";
 import Home from "./pages/Home/Home";
-import Login from "../src/pages/login/login";
-import List from "../src/pages/lists/lists";
-import EditModule from "../src/pages/EditModule/editmodule";
-import NewModule from "../src/pages/NewModule/newmod";
-import EditLecturer from "../src/pages/EditLecturer/editlecturer";
-import Newlecturer from "../src/pages/NewLecturer/newLec";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Module from "./pages/Module/Module";
+import Modules from "./pages/Modules/Modules";
 import Navbar from "./components/Navbar/Navbar";
 import Sidebar from "./components/Sidebar/Sidebar";
 import { Col, Row } from "react-bootstrap";
+import Lecturers from "./pages/Lecturers/Lecturers";
+import { useState } from "react";
 
 function App() {
+
+  const [narbar,setNavbar] = useState('Dashboard')
+
   return (
     <div className="App d-flex">
       <BrowserRouter>
         <Sidebar />
         <Col lg={2}></Col>
         <Col lg={10}>
-          <Navbar section="Modules" />
+          <Navbar section={narbar} />
           <Row className="m-0">
             <Routes>
-              <Route path="/" element={<Home />} />
-              <Route path="/modules" element={<Module />} />
+              <Route path="/" element={<Home setNavbar={setNavbar}/>} />
+              <Route path="/modules" element={<Modules setNavbar={setNavbar}/>} />
+              <Route path="/lecturers" element={<Lecturers setNavbar={setNavbar}/>} />
             </Routes>
           </Row>
         </Col>
