--- src/components/table/table.scss
@@ -1,14 +1,0 @@
-.status {
-  padding: 5px;
-  border-radius: 5px;
-
-  &.Released {
-    color: green;
-    background-color: rgba(0, 128, 0, 0.151);
-  }
-
-  &.Pending {
-    color: goldenrod;
-    background-color: rgba(189, 189, 3, 0.103);
-  }
-}
