--- src/pages/EditLecturer/editlecturer.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function editlecturer() {
-  return <div>editlecturer</div>;
-}
-
-export default editlecturer;
