--- src/pages/EditModule/editmodule.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function editmodule() {
-  return <div>editmodule</div>;
-}
-
-export default editmodule;
