--- src/pages/NewLecturer/newLec.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function newLec() {
-  return <div>newLec</div>;
-}
-
-export default newLec;
