--- src/pages/lists/lists.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function lists() {
-  return <div>lists</div>;
-}
-
-export default lists;
